ref(playstation): Move processing code to utils (like unreal)
